% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{elpa.plot.density.allyears}
\alias{elpa.plot.density.allyears}
\title{Density Plots for All Years Overlaid}
\usage{
elpa.plot.density.allyears(ds)
}
\arguments{
\item{ds}{An ELPA21 dataset}
}
\description{
Density plots show the distribution of values on a common scale.  This plot shows the density distribution of
the Scale.Score, by grade and all levels of SY (School Year) in the dataset.
}
\note{
The title is not properly adjusted if there is only one level of SY in the dataset.
}
