% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{elpa.plot.density.compare.newtodistrict}
\alias{elpa.plot.density.compare.newtodistrict}
\title{Density Distribution Comparing New To District to Students}
\usage{
elpa.plot.density.compare.newtodistrict(
  ds,
  y1,
  y2,
  overlaid = FALSE,
  label.size = 3.8
)
}
\arguments{
\item{ds}{ELPA dataset}

\item{y1}{First year to compare}

\item{y2}{Second year to compare}

\item{overlaid}{Overlay the Density Plots (TRUE) or Side-by-side (FALSE)}

\item{label.size}{Size of label on the chart}
}
\value{
A Density Plot (ggplot oblect) comparing "New to District" students to returning students.
}
\description{
A density distribution comparing students who are new to the district to students who tested in two given
years.  For this comparison, Kindergartners testing for the first time in the second year are considered
"New to District".
}
