% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{elpa.plot.density.compare}
\alias{elpa.plot.density.compare}
\title{Comparison Plot of Density for Two Years Using Z-test}
\usage{
elpa.plot.density.compare(ds, y1 = NULL, y2 = NULL)
}
\arguments{
\item{ds}{ELPA format dataset}

\item{y1}{School Year 1 for comparison.  If null, is the min of SY in the dataset.}

\item{y2}{School Year 2 for comparison.  If null, is the max of SY in the dataset.}
}
\value{
A ggplot of Density Distrbutions, colored by the Z-statistic comparison of the distributions, by grade.
}
\description{
This plot compares two distributions of the Scale.Score, by grade, using the Z-test.  The Density Distributions are
colored based on the size of the Z-statistic.
}
\details{
The Z-statistic is based on the standard normal distribution and creates a score that demonstrates the likliehood
that two distributions are from the same or different distributions.  Z-statistics less than 1.99 are considered
"Same", Z-statistics between 2 and 2.5 are considered "Marginally Different", 2.5 to 3.0 is considered
"Significantly Different", and two distributions with a Z-statistic greater than 3.0 are considered to be
"Highly Significant Different".\cr
The dataset is filtered for each identified school year and joined to compare distributions given to students in
the same grade in each year.
}
\references{
http://homework.uoregon.edu/pub/class/es202/ztest.html
}
