% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{elpa.plot.flow.building}
\alias{elpa.plot.flow.building}
\title{Flow Plot for a Single Building}
\usage{
elpa.plot.flow.building(
  df,
  school,
  y1,
  y2,
  testedyear = NULL,
  unit = "District",
  ...
)
}
\arguments{
\item{df}{An ELPA data set}

\item{school}{The name of a single school appearing in the data set.}

\item{y1}{The first year for comparison}

\item{y2}{The second year for comparison}

\item{testedyear}{(OPTIONAL) If supplied, only students who tested at the building in this year
will be included}

\item{unit}{The label for the Unit field, defaults to "District"}

\item{...}{other options to pass to elpa.plot.flow}
}
\value{
ggplot object containing an ELPA Flow Plot for a specified building
}
\description{
A version of the Flow Plot that is filtered to a single school.
}
\details{
If a *testedyear* is not supplied, then ALL students who are tested in ***either*** year at the
specified school are included in the plot.  For example, if a student tested at
School 1 in Grade 5 in Year 1, and then tested at School 2 in Grade 6 in Year 2, her scores for *both*
Grade 5 and Grade 6.  This can result in some of the counts appearing inflated.\cr\cr
If *testedyear* is supplied, then only students who actually tested at the specified building
in that year will be included.
}
