% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{elpa.plot.flow.grade}
\alias{elpa.plot.flow.grade}
\title{ELPA Flow Plot of a Single Grade}
\usage{
elpa.plot.flow.grade(df, grade, y1, y2, gradeyear = NULL, ...)
}
\arguments{
\item{df}{An ELPA dataset}

\item{grade}{The grade level to display}

\item{y1}{First year of comparison}

\item{y2}{Second year of comparison}

\item{gradeyear}{The year that the specified grade is matched.  If unspecified,
the grade is assumed to be in the Second Year.}

\item{...}{Other parameters passed elpa.plot.flow}
}
\value{
A ggplot object containing an ELPA Flow Chart
}
\description{
Creates an ELPA Flow Plot for the students in a single grade in EITHER the First Year
or the Second Year.
}
\examples{
elpa.plot.flow.grade(elpadata,'04',2021,2022,gradeyear=2022)
}
\seealso{
elpa.plot.flow
}
