% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{elpa.plot.flow}
\alias{elpa.plot.flow}
\title{Flow Plot for ELPA Students Between Two Years}
\usage{
elpa.plot.flow(
  df,
  y1,
  y2,
  adjustExit = FALSE,
  incN = FALSE,
  nsep = "\\n",
  labelflow = FALSE,
  unit = "District",
  flow.toright = TRUE
)
}
\arguments{
\item{df}{ELPA dataset}

\item{y1}{Starting Year}

\item{y2}{Ending Year}

\item{adjustExit}{(not yet implemented) Adjust "Exit" to show that some exited students may graduate or leave
the district without exiting.}

\item{incN}{TRUE/FALSE show the N sizes in the columns at the right and left}

\item{nsep}{String between labels in column at left and right and N sizes.  Defaults to \\n (newline).}

\item{labelflow}{TRUE/FALSE to put label sizes in the flows}

\item{unit}{Name of the unit.  Suggestions:  District (default), School, State}

\item{flow.toright}{TRUE Colors flows from Left to Right, highlighting movement from left to right, highlighting
movement from the Starting Year to the Ending Year.  FALSE reverses the color highlighting.}
}
\value{
ELPA Flow Plot (ggplot object)
}
\description{
A "Flow" plot (alluvial plot) showing the movement of EL students from one year to the next.
}
\note{
This package uses ONLY information available in the ELPA data set.\cr
A student is considered "New" if they only tested in the second year.\cr
Students who only tested in the first year experienced one of several possible
outcomes: (1) they may have "Left"; (2) if they were tested as Proficient in the first year, they may have
"Exited" in the second year; (3) if they were Seniors in the first year, they may have graduated at the end
of the first year and thus not tested in the second year.\cr
For convenience sake, ALL students who were proficient
when tested in the first year who did not test in the second year are counted as "Exited".\cr
The planned *adjustExit* parameter would add an adjustment to show that some proportion of the students
who tested Proficient in the first year but did not test in the second year actually remained in services
and either graduated or moved out of the district.
}
\seealso{
ggalluvial::ggalluvial
}
