% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{elpa.plot.histogram.compare.newtodistrict}
\alias{elpa.plot.histogram.compare.newtodistrict}
\title{Plot a Histogram of Scale.Score Comparing New to District to Returning Students}
\usage{
elpa.plot.histogram.compare.newtodistrict(
  ds,
  y1,
  y2,
  binwidth = 50,
  text.size.y = 10
)
}
\arguments{
\item{ds}{ELPA data set}

\item{y1}{School Year 1}

\item{y2}{School Year 2}

\item{binwidth}{Number of Scale Score points to treat as a single bar on the histogram}

\item{text.size.y}{Size of labels on the Y axis}
}
\value{
Histogram (ggplot object) of Scale Score
}
\description{
Plots a histogram of the Scale Score by levels of SY in the dataset, comparing
students who only tested in Year 2 to students who tested in both Year 1 and Year 2.
}
