% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{elpa.plot.histogram}
\alias{elpa.plot.histogram}
\title{Plot a Histogram of Scale.Score}
\usage{
elpa.plot.histogram(ds, binwidth = 50, text.size.y = 10)
}
\arguments{
\item{ds}{ELPA data set}

\item{binwidth}{Number of Scale Score points to treat as a single bar on the histogram}

\item{text.size.y}{Size of labels on the Y axis}
}
\value{
Histogram (ggplot object) of Scale Score
}
\description{
Plots a simple histogram of the Scale Score by levels of SY in the dataset.
}
