% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{elpa.plot.ntested}
\alias{elpa.plot.ntested}
\title{Plot Number of Students Tested By SY}
\usage{
elpa.plot.ntested(ds, label.size = 3)
}
\arguments{
\item{ds}{ELPA21 dataset}

\item{label.size}{Size of label, defaults to 3}
}
\value{
plot object from ggplot
}
\description{
Plots the number of students tested in each School Year (SY) included in the dataset.  Bars are split by
the overall performance level reported, and labelled with N sizes.
}
