% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{elpa.plot.profchange.bylevel.stacked}
\alias{elpa.plot.profchange.bylevel.stacked}
\title{Stacked Plot of Proficiency Change Between Two Years}
\usage{
elpa.plot.profchange.bylevel.stacked(ds, y1, y2)
}
\arguments{
\item{ds}{An ELPA data set}

\item{y1}{The first School Year to compare - defaults to min(SY) in the dataset}

\item{y2}{The second School Year for comparison - defaults to max(SY)}
}
\value{
Bar chart (ggplot) of the proportion of students who tested in both of the years specified by proficiency change.
}
\description{
A stacked version of the simple bar chart  of students who tested in both of two speficied years
showing the proportion of students by Proficiency Level in the first year compared
to proficiency level in the second year.
}
\seealso{
ggplot2::ggplot ggplot2::geom_bar ggplot2::geom_text
}
