% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{elpa.plot.testedboth}
\alias{elpa.plot.testedboth}
\title{Simple Plot of Students who Tested in Both of Two Years}
\usage{
elpa.plot.testedboth(ds, y1, y2)
}
\arguments{
\item{ds}{ELPA dataset}

\item{y1}{First Year}

\item{y2}{Second Year (must be one year after first year)}
}
\value{
Bar chart (ggplot object)
}
\description{
Growth can only be considered for students who tested in more than one year.  This bar chart shows
the number of students who tested in BOTH years and the proportion scoring each Proficiency Level by
school year.
}
