% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{fct_shift_ord}
\alias{fct_shift_ord}
\title{Shift factorial by N levels}
\usage{
fct_shift_ord(x, increment = 1, cap = TRUE, wrap = FALSE, .fun = `+`)
}
\arguments{
\item{x}{A vector of factors to be shifted}

\item{increment}{The number of levels to shift by applying function .fun (defaults to add)}

\item{cap}{If true, return first element if factor level shifted below first level
or last level if shifted above last level; otherwise return NA if level shifts out
of bounds}

\item{wrap}{UNIMPLEMENTED, if TRUE, wrap values around when shifting out of bounds;
ie. next level of last level is first level.}

\item{.fun}{Function to apply when shifting levels.  Default is `+` to increment to
next level, `-` will shift to previous level, other functions will work but may not
be meaningful}
}
\value{
The shifted value
}
\description{
Function
}
