% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{findnewestfile}
\alias{findnewestfile}
\title{Find the newest file (Windows only)}
\usage{
findnewestfile(pattern, read.data = TRUE, path = NULL, ...)
}
\arguments{
\item{pattern}{A grep-style pattern (NOT Windows wildcards) that is used to find the file}

\item{read.data}{logical - TRUE include the contents of the file in the list}

\item{path}{A system-style path value}

\item{...}{Other options passed to read.csv}
}
\value{
LIST with members path, created, age, data
}
\description{
Searches a path for a file matching the pattern and returns a list of items, including the contents
of the file, for common file parts
}
