% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{helper.buildelpadata}
\alias{helper.buildelpadata}
\title{Build a sample data set to include with the package}
\usage{
helper.buildelpadata()
}
\description{
This function is used to generate the "elpadata" dataset that will be incldued with the
completed package.
}
\note{
This function depends on files that are located on the developer's system, and are not
included in the package.
}
\seealso{
readelpascores, helper.deidentify
}
