% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{helper.deidentify}
\alias{helper.deidentify}
\title{Replace Identifying Information With Randomized Values}
\usage{
helper.deidentify(
  ds,
  idvar = .data$Student.ID,
  prefix = NULL,
  sample.size = NULL
)
}
\arguments{
\item{ds}{A dataframe or tibble.}

\item{idvar}{The variable to replace}

\item{prefix}{(Optional) A prefix to place in front of the identifier, default is two random lowercase letters.}

\item{sample.size}{(Optional) If supplied, return a dataset with only N levels of the idvar included}
}
\value{
In final form, a dataframe in the same format with the contents of the idvar column replaced.
}
\description{
This function replaces Personally Identifiable Information with randomly ordered numeric identifiers between
1 and the number of distinct values for the identifier in the dataset.
}
\details{
If an identifier is repeated in the data set, it is replaced in all occurrences with its intended replacement.
The numeric portion of the replacement identifier is padded with zeros to ensure that all identifiers are the
same length.\cr\cr
If a sample.size is specified, the returned data set will (usually) contain more than sample.size rows; however,
there will only be sample.size distinct values for the idvar in the resulting dataset.
}
