% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{helper.ethnicity.groups}
\alias{helper.ethnicity.groups}
\title{Collapses ethnicity to three groups}
\usage{
helper.ethnicity.groups(Ethnicity)
}
\arguments{
\item{Ethnicity}{Ethnicity, as modified by \code{helper.cleandata()}}
}
\value{
Grouped ethnicity, as factor
}
\description{
Returns Ethnicity collapsed for charts
}
\details{
In Springdale, the largest groups of language learners are Hispanic, followed by Pacific Islanders.
This function returns ethnicity as simply 'Hispanic', 'Islander' or 'Other'.
}
