% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{helper.gradeband}
\alias{helper.gradeband}
\title{Map grades to ELPA21 Grade Bands}
\usage{
helper.gradeband(grade)
}
\arguments{
\item{grade}{Grade number}
}
\value{
Ordered factor for gradeband
}
\description{
The ELPA21 test is differentiated across Grade Bands; this helper function returns
a grade band for the grade passed.
}
\details{
Built with bands as of 2019: KG, 1, 2-3, 4-5, 6-8, 9-12.  Unknown values return
Grade Band 9-12.  Value is returned as an ordered factor.
}
