% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{helper.isvalidelpa}
\alias{helper.isvalidelpa}
\title{Is the data set formatted as an ELPA data set?}
\usage{
helper.isvalidelpa(ds = NULL)
}
\arguments{
\item{ds}{A dataframe to validate}
}
\value{
TRUE or FALSE
}
\description{
A simple helper function that just checks for valid variable names in the dataset.  Will error if there are any
extra variables.
}
\details{
Expects to find these variables: Student.ID, School, Ethnicity, Scale.Score, Performance.Level, SY
}
