% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{helper.reducedata}
\alias{helper.reducedata}
\title{Reduce ELPA dataset}
\usage{
helper.reducedata(ds, SY = 2022)
}
\arguments{
\item{ds}{A dataset read by the \code{readelpadata()} function}

\item{SY}{The School Year for the dataset}
}
\value{
data frame formatted as ELPA data for this package
}
\description{
Reduce a data set read using the \code{readelpadata()} function to the format expected by
\code{helper.isvalidelpa()} function.
}
