% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{helper.schools.with.grade}
\alias{helper.schools.with.grade}
\title{Returns an alphabetized list of schools that serve a specified grade or grades}
\usage{
helper.schools.with.grade(df, grade)
}
\arguments{
\item{df}{ELPA data set}

\item{grade}{a single grade or a vector of grades}
}
\value{
Schools that serve the grade(s) indicated ordered alphabetically, as a vector of character strings
}
\description{
Identify schools that serve a specified grade *or* any one of the grades in a vector of grades
}
\examples{
helper.schools.with.grade(elpadata,'KG')
helper.schools.with.grade(elpadata,c('KG','06'))
}
