% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{helper.titlepage}
\alias{helper.titlepage}
\title{Creates a Title Page using ggplot}
\usage{
helper.titlepage(
  title = "",
  subtitle = "",
  author = "",
  footnote = "",
  cornertext = "",
  textbox = "",
  textbox.grow = FALSE,
  title.face = "bold",
  title.family = "serif",
  title.reflow = TRUE,
  subtitle.face = "plain",
  subtitle.reflow = FALSE,
  author.face = "italic"
)
}
\arguments{
\item{title}{String to place as a title block on the page}

\item{subtitle}{A slightly smaller second line of a title}

\item{author}{String positioned immediately below the subtitle}

\item{footnote}{A string of text that is positioned at the lower right corner of the page}

\item{cornertext}{String to place in the top left corner of the page}

\item{textbox}{A string to place as a textbox in the "middle" of the title page}

\item{textbox.grow}{TRUE - grow the text in the textbox to fill the allowed space, default=FALSE}

\item{title.face}{Font face for the title (plain, bold, italic, etc.)}

\item{title.family}{Font family for the title}

\item{title.reflow}{TRUE - allow the Title to be reflown across multiple lines}

\item{subtitle.face}{Font face for the subtitle (plain, bold, italic, etc.)}

\item{subtitle.reflow}{TRUE - allow the Subtitle to reflown across multiple lines}

\item{author.face}{Font face for the "author" line (plain, bold, italic, etc.)}
}
\value{
ggplot object that contains a title page
}
\description{
This function creates a simple title page that can be included in a PDF or set of graphical slides
}
\details{
Uses ggplot and geom_fit_text to create a title page with standardized options.
}
\examples{
helper.titlepage("ELPA Graphics in R","A Package to Help Explore ELPA21 Data","Albert Crosby")
}
