% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{readelpascores}
\alias{readelpascores}
\title{Reads a directory of ELPA scores}
\usage{
readelpascores(elpafiledirectory = "./")
}
\arguments{
\item{elpafiledirectory}{The path to the location where the ELPA21 files have been expanded.}
}
\value{
ELPA Dataset data.frame
}
\description{
Reads the contents of a directory where the ELPA21 scores from the file portal have been expanded.
User is expected to download the CSV files from the ELPA21 portal, and unzip them.
}
\note{
The tool expects files with names as downloaded from the ELPA21 Portal.  The files
should all be of type CSV.  The naming convention should be <School>_<TestName>_<FileType>_<Date>.CSV
where <TestName> includes the grade level in  the format "Kindergarten" or "Grade1" through "Grade12".
}
