% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{elpa.plot.proflevel.byyears}
\alias{elpa.plot.proflevel.byyears}
\title{Plot Proficiency Change Across Years}
\usage{
elpa.plot.proflevel.byyears(
  ds,
  SY.labels = NULL,
  SY.labels.size = 10,
  move.size = 10,
  move.labels = NULL,
  ntested = TRUE,
  ntested.size = 3
)
}
\arguments{
\item{ds}{A valid ELPA dataset}

\item{SY.labels}{The labels for the school years on the X axis}

\item{SY.labels.size}{The size for the X axis labels}

\item{move.size}{The max size for the "movement" labels placed between the columns}

\item{move.labels}{The labels for "UP"/"DOWN" in the middle.}

\item{ntested}{TRUE (default) include number of students tested on chart as part of X axis
labels, will not work properly with faceted charts}

\item{ntested.size}{The size for the N Tested label at the bottom of the chart}
}
\value{
ggplot2 grob
}
\description{
This plot creates a series of columns for proficiency by years, labelled with counts and
pecentages on each column, and the space between columns labelled with the percentage
change in distribution from year 1 to year 2.
}
