% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-ELPA21.R
\name{readelpascores}
\alias{readelpascores}
\title{Reads a directory of ELPA scores}
\usage{
readelpascores(
  elpafiledirectory = "C:/Users/acrosby/Downloads/ELPASpring2022/"
)
}
\arguments{
\item{elpafiledirectory}{The path to the location where the ELPA21 files have been expanded.}
}
\value{
ELPA Dataset data.frame
}
\description{
Reads the contents of a directory where the ELPA21 scores from the file portal have been expanded.
User is expected to download the CSV files from the ELPA21 portal, and unzip them.
}
\note{
This helper package assumes that there are 13 grades for your district - K through 12.
For districts/schools with fewer grades, errors may result.  The tool needs testing to see if
it can be made to work correctly with a partial dataset.
}
